Extract controller configs into controller_configs.py, clean up imports
- Move DK1ControllerConfig, PosVelControllerConfig, TorquePosControllerConfig,
  and MITControllerConfig out of follower.py into controller_configs.py
- Fix bi_follower.py to import from controller_configs instead of follower
- Add MITControllerConfig to bi_follower.py imports
- Remove unused ensure_safe_goal_position import from follower.py
- Lower requires-python to >=3.10 for Jetson compatibility
