+++ pyproject.toml
@@ -5,7 +5,7 @@ build-backend = "hatchling.build"
 [project]
 name = "lerobot_robot_trlc_dk1"
 version = "0.3.0"
-requires-python = ">=3.12"
+requires-python = ">=3.10"
 dependencies = [
   "numpy>=2.0.1",
   "pyserial>=3.5",
